scripts/requirements.txt
# Requirements for sign_exclusion_list.py
base58==2.1.1
PyNaCl==1.5.0